feat: Add DataTableViewOptions and DataTableRowActions components
Add new components DataTableViewOptions and DataTableRowActions to provide additional functionality to the data table. DataTableViewOptions component adds a dropdown menu for toggling column visibility in the table. DataTableRowActions component adds a dropdown menu for performing actions on individual rows in the table. These components enhance the user experience and improve the usability of the data table.